# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BarbrGo - Barber Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}/bookings:
    get:
      operationId: getUserBookings
      tags: [users]
      summary: Get user bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

  # Barbers
...
  /barbers/{id}/bookings:
    get:
      operationId: getBarberBookings
      tags: [barbers]
      summary: Get barber bookings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Barber bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /barbers/{id}/reviews:
    get:
      operationId: getBarberReviews
      tags: [barbers]
      summary: Get barber reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # Reviews
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Offers
  /offers:
    get:
      operationId: getOffers
      tags: [offers]
      summary: Get active offers
      responses:
        "200":
          description: Active offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/validate:
    post:
      operationId: validateCoupon
      tags: [offers]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponRequest"
      responses:
        "200":
          description: Coupon validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResponse"

  # Admin
...
  /admin/bookings:
    get:
      operationId: getAdminBookings
      tags: [admin]
      summary: Get all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingsResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Service:
      type: object
      properties:
        id:
          type: integer
        barberId:
          type: integer
        serviceTypeId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        description:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - barberId
        - name
        - price
        - durationMinutes
        - isActive

    # Bookings
    CreateBookingRequest:
      type: object
      properties:
        userId:
          type: integer
        barberId:
          type: integer
        serviceIds:
          type: array
          items:
            type: integer
        date:
          type: string
        startTime:
          type: string
        isHomeService:
          type: boolean
        address:
          type: string
        paymentMethod:
          type: string
          enum: [online, cash]
        couponCode:
          type: string
        notes:
          type: string
      required:
        - userId
        - barberId
        - serviceIds
        - date
        - startTime
        - paymentMethod
    UpdateBookingRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
    CancelBookingRequest:
      type: object
      properties:
        reason:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        barberId:
          type: integer
        barberName:
          type: string
        barberAvatar:
          type: string
        userName:
          type: string
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalDuration:
          type: integer
        totalAmount:
          type: number
        discountAmount:
          type: number
        finalAmount:
          type: number
        isHomeService:
          type: boolean
        address:
          type: string
        paymentMethod:
          type: string
          enum: [online, cash]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        couponCode:
          type: string
        notes:
          type: string
        cancelReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - barberId
        - barberName
        - userName
        - services
        - date
        - startTime
        - endTime
        - totalDuration
        - totalAmount
        - discountAmount
        - finalAmount
        - isHomeService
        - paymentMethod
        - paymentStatus
        - status
        - createdAt

    # Reviews
    CreateReviewRequest:
      type: object
      properties:
        bookingId:
          type: integer
        barberId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - bookingId
        - barberId
        - rating
    Review:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        barberId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - barberId
        - userId
        - userName
        - rating
        - createdAt

    # Offers
    CreateOfferRequest:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minBookingAmount:
          type: number
        maxDiscount:
          type: number
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        maxUsage:
          type: integer
        isFirstTimeOnly:
          type: boolean
      required:
        - code
        - description
        - discountType
        - discountValue
    Offer:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minBookingAmount:
          type: number
        maxDiscount:
          type: number
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        maxUsage:
          type: integer
        currentUsage:
          type: integer
        isFirstTimeOnly:
          type: boolean
        isActive:
          type: boolean
      required:
        - id
        - code
        - description
        - discountType
        - discountValue
        - isActive
    ValidateCouponRequest:
      type: object
      properties:
        code:
          type: string
        userId:
          type: integer
        amount:
          type: number
      required:
        - code
        - userId
        - amount
    CouponValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        discountAmount:
          type: number
        message:
          type: string
      required:
        - valid
        - message

    # Admin
...
    AdminBookingsResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
      required:
        - bookings
        - total
...